+++ example/gobatch.jcl
@@ -6,7 +6,7 @@
 // PARM=('POSIX(ON) FILETAG(AUTOCVT,AUTOTAG)/')
 //STEPLIB DD DSN=&SYSUID..GO.PDSE,DISP=SHARE
 //SYSPRINT DD SYSOUT=*,DCB=(RECFM=V,LRECL=133)
-//SYSOUT  DD SYSOUT=*,DCB=(RECFM=V,LRECL=133)
+//SYSOUT DD SYSOUT=*,DCB=(RECFM=V,LRECL=133)
 //SYSTERM DD SYSOUT=*,DCB=(RECFM=V,LRECL=133)
 //SYSUDUMP DD DSN=&SYSUID..DUMPS,DISP=SHARE
 //SYSMDUMP DD DSN=&SYSUID..DUMPS,DISP=SHARE
