+++ record.go
@@ -7,7 +7,6 @@
 package zosrecordio
 
 import (
-	"reflect"
 	"runtime"
 	"syscall"
 	"unsafe"
@@ -25,22 +24,24 @@ func (rs RecordStream) Nil() bool {
 	return rs.s == 0
 }
 
-// Fopen takes a name and mode as byte slices and returns a RecordStream.
-// Note that the strings encoded in thee slices must be null terminated.
-// The utility function ConvertStringToSlice automatically null-terminates.
-func Fopen(fname []byte, mode []byte) (rs RecordStream) {
+// Fopen takes a name and mode as strings  and returns a RecordStream.
+func Fopen(fname string, mode string) (rs RecordStream) {
+	fnameBytes := []byte(fname + "\x00")
+	modeBytes := []byte(mode + "\x00")
 	ret := runtime.CallLeFuncByPtr(runtime.XplinkLibvec+0x753<<4, //fopen
-		[]uintptr{uintptr(unsafe.Pointer(&fname[0])),
-			uintptr(unsafe.Pointer(&mode[0]))})
+		[]uintptr{uintptr(unsafe.Pointer(&fnameBytes[0])),
+			uintptr(unsafe.Pointer(&modeBytes[0]))})
 	rs.s = ret
 	return rs
 }
 
 // Freopen behaves the same as Fopen, but takes a previously used RecordStream
-func Freopen(fname []byte, mode []byte, rs RecordStream) (rso RecordStream) {
+func Freopen(fname string, mode string, rs RecordStream) (rso RecordStream) {
+	fnameBytes := []byte(fname + "\x00")
+	modeBytes := []byte(mode + "\x00")
 	ret := runtime.CallLeFuncByPtr(runtime.XplinkLibvec+0x754<<4, //freopen
-		[]uintptr{uintptr(unsafe.Pointer(&fname[0])),
-			uintptr(unsafe.Pointer(&mode[0])),
+		[]uintptr{uintptr(unsafe.Pointer(&fnameBytes[0])),
+			uintptr(unsafe.Pointer(&modeBytes[0])),
 			rs.s})
 	rso.s = ret
 	return rso
@@ -121,54 +122,6 @@ func (rs RecordStream) Fclose() int {
 	ret := runtime.CallLeFuncByPtr(runtime.XplinkLibvec+0x067<<4, //fclose
 		[]uintptr{rs.s})
 	return int(ret)
-}
-
-// ConvertStringToSlice copies the string into the given slice.
-// Always includeds a null terminator in the copy.
-// Returns a new empty slice if the string doesn't fit.
-func ConvertStringToSlice(s string, bi []byte) (bo []byte) {
-	size := len(s)
-	if size < cap(bi) {
-		copy(bi[:size], s)
-		bi[size] = 0
-		return bi[:size+1]
-	} else {
-		return []byte{}
-	}
-}
-
-// ConvertStructToSlice returns a byte slice that shares storage
-// with the incoming struct. The length of the slice will be
-// exactly the size of the struct.
-func ConvertStructToSlice(i interface{}) (slice []byte) {
-	var size int
-	var ptr unsafe.Pointer
-	ptr = unsafe.Pointer(reflect.ValueOf(i).Elem().UnsafeAddr())
-	size = int(reflect.ValueOf(i).Elem().Type().Size())
-	data := (*(*[1<<31 - 1]byte)(ptr))[:size]
-	return data
-}
-
-// ConvertSliceToStruc returns an interface that can be type asserted to be the
-// same type as the incoming pointer to struct "i".
-// The pointer that results from such a type assertion will share storage
-// with the incoming byte slice "bi". The struct size is the second returned value.
-// Thus the following sequence leaves buffp as a pointer to a FixedHeader struct
-// that shares storage with myBigSlice[:buffSize]:
-// var buffp *FixedHeader_T
-// buffp_, buffSize := zosrecordio.ConvertSliceToStruct(buffp, myBigSlice)
-// buffp = buffp_.(*FixedHeader_T)
-// Note: if the  incoming slice isn't big enough, it returns <nil, 0>.
-func ConvertSliceToStruct(i interface{}, bi []byte) (interface{}, int) {
-	bip := (&bi[0])
-	bipp := &bip
-	result := reflect.NewAt(reflect.ValueOf(i).Type(), unsafe.Pointer(bipp)).Elem().Interface()
-	size := int(reflect.ValueOf(result).Elem().Type().Size())
-	if size > len(bi) {
-		return nil, 0
-	} else {
-		return result, size
-	}
 }
 
 // The equivalent of STDIN
